:recycle: Moved survey controller unit test to its appropriate package.
